Move rule and its formatters to another file
